# beep-backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Room Booking Application API",
    "description": "Documentation for the API",
    "license": {
      "name": " ",
      "url": " "
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/reservation/": {
      "post": {
        "tags": [
          "Reservation"
        ],
        "description": "Create new Reservation in system",
        "parameters": [
          {
            "name": "reservation",
            "in": "body",
            "description": "Reservation that we want to create",
            "schema": {
              "$ref": "#/definitions/Reservation"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New Reservation was created",
            "schema": {
              "$ref": "#/definitions/Reservation"
            }
          }
        }
      }
    },
    "/api/reservation/{id_room}": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "description": "get a Reservation from data base",
        "parameters": [
          {
            "name": "id_room",
            "in": "params",
            "description": "get reservation by ID"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New Reservation was created",
            "schema": {
              "$ref": "#/definitions/Reservation"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Reservation": {
      "required": [
        "id_room",
        "start_date",
        "end_date"
      ],
      "properties": {
        "id_room": {
          "type": "string",
          "uniqueItems": true
        },
        "start_date": {
          "type": "string"
        },
        "end_date": {
          "type": "string"
        }
      }
    }
  }
}